Safari-609.1.11.1/LayoutTests/legacy-animation-engine/animations/animation-events-not-cancelable-expected.txt

Tests that animation events are not cancelable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


animationstart:
PASS event.cancelable is false
animationiteration:
PASS event.cancelable is false
animationend:
PASS event.cancelable is false
PASS successfullyParsed is true

TEST COMPLETE
